requirements.txt
paho-mqtt
sqlalchemy
redis
psycopg2-binary
pyserial
pytz
